Adding 'edit' state inside logic, graphic and GUI - This new state is used in editor and not the 'play' state anymore

+++ intern/src/editor/edit_application.cpp
@@ -190,7 +190,7 @@ namespace
             // -----------------------------------------------------------------------------
             Edit::CMessage NewMessage;
 
-            NewMessage.PutFloat(Core::Time::GetDeltaTimeLastFrame());
+            NewMessage.PutDouble(Core::Time::GetDeltaTimeLastFrame());
 
             NewMessage.Reset();
